refactor(types): annotate model casts (@property) for enums/dates/arrays
Larastan does not infer the casts() enum/date/array casts, so magic-property
access fell back to string. Add @property docblocks for status (enum),
ownership, date columns (Carbon) and array columns on Album/Game/Anime/Comic/Book.
This clears the form-tail errors (status->value, date->format(), is_array on
genre) and tidies BookSettings cover recaching.

Baseline: 495 -> 466 errors, zero additions.

+++ app/Livewire/Books/BookForm.php
@@ -226,7 +226,7 @@ class BookForm extends Component
         return view('livewire.books.book-form', [
             'statuses' => $this->getStatuses(),
             'isEditing' => $this->isEditing(),
-            'availableTags' => Book::getAllTagsForUser(Auth::id()),
+            'availableTags' => Book::getAllTagsForUser((int) Auth::id()),
         ]);
     }
 }
